requirements.txt
PyYAML==6.0.0
pytest==7.4.2
jsonschema==4.19.1
mock==5.1.0
setuptools~=68.2.0
pandas
faker